# api/openapi.yaml
openapi: 3.1.0
info:
  title: Language Learning Platform API
  description: |
    Central API contract for the language learning platform.

    **Naming conventions**
    - API JSON / OpenAPI: `snake_case` (e.g. `lesson_id`, `expected_answer`)
    - Java entity fields: `camelCase` (e.g. `lessonId`, `expectedAnswer`)
    - Database columns: `snake_case` via `@Column(name = ...)`

    Service ownership is indicated per-operation via `x-service`.
  version: 0.1.0
servers:
  - url: http://localhost:8081
    description: User Service
  - url: http://localhost:8082
    description: Learning Service
  - url: http://localhost:8083
    description: Progress-Feedback Service
  - url: http://localhost:8084
    description: GenAI Service
tags:
  - name: exercises
    x-displayName: exercises
  - name: writing
    x-displayName: writing
  - name: speaking
    x-displayName: speaking
  - name: ops
    x-displayName: ops
paths:
  /api/v1/genai/exercises/generate:
    post:
      tags:
        - exercises
      summary: Generate additional exercises
      description: >-
        Given a lesson topic and a set of existing exercises as style
        references, generate new exercises using AI. The AI matches the format,
        difficulty, and exercise types of the provided reference set.
      operationId: generateExercises
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateExercisesRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateExercisesResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-service: genai-service
  /api/v1/genai/writing/evaluate:
    post:
      tags:
        - writing
      summary: Evaluate a written answer
      description: >-
        Evaluate a learner's written answer against the exercise's expected
        answer. Returns a numeric score (0–10) ready to store as
        UserAnswer.score, plus message and weak_area fields that map directly to
        a Feedback record. Also returns the corrected answer and a detailed
        explanation of errors.
      operationId: evaluateWriting
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WritingEvaluationRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WritingEvaluationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-service: genai-service
  /api/v1/genai/speaking/evaluate:
    post:
      tags:
        - speaking
      summary: Evaluate a spoken answer
      description: >-
        Transcribe the learner's audio with Whisper, then evaluate the
        transcription against the expected answer using the configured LLM.
        Returns a numeric score (0–10), feedback, the corrected answer, and
        optionally a base64-encoded WAV of the corrected answer spoken aloud
        (when TTS is enabled).
      operationId: evaluateSpeaking
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_evaluateSpeaking'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeakingEvaluationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-service: genai-service
  /api/v1/genai/speaking/practice:
    post:
      tags:
        - speaking
      summary: Conversational speaking practice
      description: >-
        One turn of a free-form spoken conversation with an AI character in the
        target language. The frontend maintains the conversation history and
        sends it with each request. On the final turn (end_session=true) the
        service additionally returns a list of language corrections made across
        the full session, plus a spoken audio summary.
      operationId: speakingPractice
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_speakingPractice'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeakingPracticeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-service: genai-service
  /health:
    get:
      tags:
        - ops
      summary: Health
      operationId: health_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
components:
  schemas:
    Body_evaluateSpeaking:
      properties:
        audio:
          type: string
          contentMediaType: application/octet-stream
          title: Audio
          description: Audio recording of the learner's answer (WAV, MP3, WebM, etc.)
        user_id:
          type: integer
          title: User Id
        exercise_id:
          type: integer
          title: Exercise Id
        exercise_type:
          type: string
          title: Exercise Type
          examples:
            - translation
        question:
          type: string
          title: Question
          examples:
            - 'Translate: ''The cat is on the table'''
        expected_answer:
          type: string
          title: Expected Answer
          examples:
            - Die Katze ist auf dem Tisch
        target_language:
          type: string
          title: Target Language
          examples:
            - German
        level:
          type: string
          title: Level
          examples:
            - A2
      type: object
      required:
        - audio
        - user_id
        - exercise_id
        - exercise_type
        - question
        - expected_answer
        - target_language
        - level
      title: Body_evaluateSpeaking
    Body_speakingPractice:
      properties:
        audio:
          type: string
          contentMediaType: application/octet-stream
          title: Audio
          description: Audio of the learner's spoken turn
        scenario:
          type: string
          title: Scenario
          examples:
            - waiter at a German restaurant
        target_language:
          type: string
          title: Target Language
          examples:
            - German
        level:
          type: string
          title: Level
          examples:
            - A2
        history_json:
          type: string
          title: History Json
          description: >-
            JSON-serialised list of ConversationMessage objects from previous
            turns
          default: '[]'
        end_session:
          type: boolean
          title: End Session
          description: Set to true on the final turn to trigger end-of-session corrections
          default: false
      type: object
      required:
        - audio
        - scenario
        - target_language
        - level
      title: Body_speakingPractice
    ExerciseContext:
      properties:
        type:
          type: string
          title: Type
          description: >-
            Exercise type, e.g. 'translation', 'fill-in-the-blank',
            'multiple-choice'
          examples:
            - translation
        question:
          type: string
          title: Question
          description: The exercise prompt shown to the learner
          examples:
            - 'Translate: ''The cat is on the table'''
        difficulty:
          type: string
          title: Difficulty
          examples:
            - beginner
        expected_answer:
          anyOf:
            - type: string
            - type: 'null'
          title: Expected Answer
          examples:
            - Die Katze ist auf dem Tisch
      type: object
      required:
        - type
        - question
        - difficulty
      title: ExerciseContext
      description: An existing exercise provided as a style/difficulty reference.
    GenerateExercisesRequest:
      properties:
        lesson_id:
          type: integer
          title: Lesson Id
          description: >-
            ID of the lesson these exercises belong to — passed through to each
            generated exercise
          examples:
            - 3
        lesson_topic:
          type: string
          title: Lesson Topic
          description: Topic or theme of the lesson
          examples:
            - Everyday household items
        target_language:
          type: string
          title: Target Language
          description: Language being learned (full name or BCP-47 tag)
          examples:
            - German
        level:
          type: string
          title: Level
          description: CEFR level or descriptive level of the learner
          examples:
            - A2
        existing_exercises:
          items:
            $ref: '#/components/schemas/ExerciseContext'
          type: array
          maxItems: 5
          minItems: 1
          title: Existing Exercises
          description: 1–5 existing exercises to use as style and difficulty references
        count:
          type: integer
          maximum: 10
          minimum: 1
          title: Count
          description: Number of new exercises to generate
          default: 3
          examples:
            - 3
      type: object
      required:
        - lesson_id
        - lesson_topic
        - target_language
        - level
        - existing_exercises
      title: GenerateExercisesRequest
      description: >-
        Request body for generating additional exercises similar to existing
        ones.

        Provide 1–5 existing exercises as style references — the AI will match
        their

        format, difficulty, and language level when creating new ones.
    GenerateExercisesResponse:
      properties:
        exercises:
          items:
            $ref: '#/components/schemas/GeneratedExercise'
          type: array
          title: Exercises
          description: Newly generated exercises in the same style as the provided examples
      type: object
      required:
        - exercises
      title: GenerateExercisesResponse
      description: >-
        Response body containing newly generated exercises.

        Each exercise matches the style and difficulty of the provided reference
        set.
    GeneratedExercise:
      properties:
        lesson_id:
          type: integer
          title: Lesson Id
          description: Lesson this exercise belongs to — passed through from the request
        type:
          type: string
          enum:
            - translation
            - fill-in-the-blank
            - multiple-choice
            - sentence-building
          title: Type
          description: Exercise type — maps to Exercise.type
        question:
          type: string
          title: Question
          description: The exercise prompt — maps to Exercise.question
        difficulty:
          type: string
          title: Difficulty
          description: Difficulty level — maps to Exercise.difficulty
        expected_answer:
          type: string
          title: Expected Answer
          description: The correct answer — maps to Exercise.expectedAnswer
      type: object
      required:
        - lesson_id
        - type
        - question
        - difficulty
        - expected_answer
      title: GeneratedExercise
      description: >-
        A single AI-generated exercise. All fields map directly to Exercise in
        the backend.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    SpeakingEvaluationResponse:
      properties:
        transcription:
          type: string
          title: Transcription
          description: Speech-to-text transcription of the learner's audio
          examples:
            - Die Katze ist an den Tisch
        score:
          type: number
          maximum: 10
          minimum: 0
          title: Score
          description: Numeric score 0–10 — maps to UserAnswer.score
          examples:
            - 6.5
        is_correct:
          type: boolean
          title: Is Correct
          description: True if the spoken answer is substantially correct
        message:
          type: string
          title: Message
          description: >-
            Full feedback covering what was wrong and why — maps to
            Feedback.message. Should be a complete, helpful explanation suitable
            for storing as the feedback record.
          examples:
            - >-
              Good attempt! You used 'an' (contact/vertical surface) instead of
              'auf' (horizontal surface). Also check the case: 'Tisch' requires
              the dative 'dem' here.
        weak_area:
          type: string
          title: Weak Area
          description: Short category to improve — maps to Feedback.weakArea
          examples:
            - preposition usage
        corrected_answer:
          type: string
          title: Corrected Answer
          description: The corrected version of the learner's answer — for frontend display
          examples:
            - Die Katze ist auf dem Tisch
        feedback_audio_b64:
          anyOf:
            - type: string
            - type: 'null'
          title: Feedback Audio B64
          description: >-
            Base64-encoded WAV audio of corrected_answer spoken aloud (null when
            TTS is disabled)
      type: object
      required:
        - transcription
        - score
        - is_correct
        - message
        - weak_area
        - corrected_answer
      title: SpeakingEvaluationResponse
      description: >-
        AI evaluation of a spoken answer.

        score and feedback fields mirror WritingEvaluationResponse so both map
        to

        the same UserAnswer + Feedback domain objects in the backend.

        transcription is the raw Whisper output — useful for displaying what was
        heard.

        feedback_audio_b64 is a base64-encoded WAV of the corrected answer
        spoken aloud

        (only present when TTS is enabled).
    SpeakingPracticeResponse:
      properties:
        transcription:
          type: string
          title: Transcription
          description: Whisper transcription of the learner's audio for this turn
        ai_response_text:
          type: string
          title: Ai Response Text
          description: AI character's reply in the target language
        ai_response_audio_b64:
          anyOf:
            - type: string
            - type: 'null'
          title: Ai Response Audio B64
          description: Base64-encoded WAV of the AI's reply spoken aloud
        session_corrections:
          anyOf:
            - items:
                $ref: '#/components/schemas/TurnCorrection'
              type: array
            - type: 'null'
          title: Session Corrections
          description: >-
            Language errors found across the full session — only present on the
            final turn
        corrections_audio_b64:
          anyOf:
            - type: string
            - type: 'null'
          title: Corrections Audio B64
          description: >-
            Base64-encoded WAV of the corrections summary spoken aloud — only on
            the final turn
      type: object
      required:
        - transcription
        - ai_response_text
      title: SpeakingPracticeResponse
      description: >-
        Response for one turn of a speaking practice conversation.

        On normal turns only transcription + AI reply are populated.

        On the final turn (end_session=true) session_corrections and
        corrections_audio_b64

        are also populated.
    TurnCorrection:
      properties:
        original:
          type: string
          title: Original
          description: What the learner actually said
          examples:
            - Ich möchte ein Kaffee
        corrected:
          type: string
          title: Corrected
          description: What they should have said
          examples:
            - Ich möchte einen Kaffee
        explanation:
          type: string
          title: Explanation
          description: Short grammar or vocabulary note
          examples:
            - >-
              'Kaffee' is masculine, so the accusative article is 'einen' not
              'ein'
      type: object
      required:
        - original
        - corrected
        - explanation
      title: TurnCorrection
      description: A single language error identified across the session.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
    WritingEvaluationRequest:
      properties:
        user_id:
          type: integer
          title: User Id
          examples:
            - 42
        exercise_id:
          type: integer
          title: Exercise Id
          examples:
            - 7
        exercise_type:
          type: string
          title: Exercise Type
          description: Type of the exercise being answered
          examples:
            - translation
        question:
          type: string
          title: Question
          description: The original exercise question shown to the learner
          examples:
            - 'Translate: ''I would like a coffee, please'''
        expected_answer:
          type: string
          title: Expected Answer
          description: The canonical correct answer
          examples:
            - Je voudrais un café, s'il vous plaît
        user_answer:
          type: string
          title: User Answer
          description: The learner's submitted written answer
          examples:
            - Je voudrai un cafe s'il vous plait
        target_language:
          type: string
          title: Target Language
          examples:
            - French
        level:
          type: string
          title: Level
          examples:
            - A2
      type: object
      required:
        - user_id
        - exercise_id
        - exercise_type
        - question
        - expected_answer
        - user_answer
        - target_language
        - level
      title: WritingEvaluationRequest
      description: >-
        Request body for evaluating a learner's written answer against an
        exercise.

        Maps directly to the UserAnswer + Exercise domain objects in the Java
        backend.
    WritingEvaluationResponse:
      properties:
        score:
          type: number
          maximum: 10
          minimum: 0
          title: Score
          description: Numeric score 0–10 — maps to UserAnswer.score
          examples:
            - 6.5
        is_correct:
          type: boolean
          title: Is Correct
          description: True if the answer is substantially correct
        message:
          type: string
          title: Message
          description: >-
            Full feedback including what was wrong and why — maps to
            Feedback.message. Should be a complete, helpful explanation suitable
            for storing as the feedback record.
          examples:
            - >-
              Good attempt! 'voudrais' needs a conditional ending (-ais not
              -ai), and French requires accents: 'café' not 'cafe'. Otherwise
              the structure is correct.
        weak_area:
          type: string
          title: Weak Area
          description: >-
            Short grammatical/lexical category to improve — maps to
            Feedback.weakArea
          examples:
            - accents and diacritics
        corrected_answer:
          type: string
          title: Corrected Answer
          description: The corrected version of the learner's answer — for frontend display
          examples:
            - Je voudrais un café, s'il vous plaît
      type: object
      required:
        - score
        - is_correct
        - message
        - weak_area
        - corrected_answer
      title: WritingEvaluationResponse
      description: >-
        AI evaluation of a written answer.

        score maps to UserAnswer.score; message and weak_area map directly to

        the Feedback entity (Feedback.message, Feedback.weakArea) in the
        backend.
x-tagGroups:
  - name: GenAI Language Learning Service
    tags:
      - exercises
      - writing
      - speaking
      - ops